requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning
scikit-learn>=1.3.0

# Google Gemini AI
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
datetime  # Built-in, no need to install

# Additional useful libraries (opcional pero recomendado)
openpyxl>=3.1.0  # Para leer archivos Excel
xlrd>=2.0.0      # Para archivos Excel antiguos